requirements.txt
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
motor
python-dotenv
httpx
deepface
tensorflow
tf-keras
opencv-python-headless
numpy
sqlalchemy
bcrypt
opencv-python
pandas
python-multipart
matplotlib
tqdm
aiofiles
pydub
slowapi
vaderSentiment
decouple